Refactor ID types and improve exception handling
chore Updated AutoMapperProfile and BookingDTO to use Guid? for IDs.
Removed unused ItemGroup from TravelBooking.Common.csproj.
Enhanced ExceptionHandlingMiddleware to handle ValidationExceptionList.
Moved ValidationExceptionList to a new file in TravelBooking.Common.Exceptions.
Updated repositories to throw KeyNotFoundException for not found entities.
